Fixes LL-3757
