+++ src/families/ethereum/modules/compound.js
@@ -22,15 +22,12 @@ import type { ModeModule } from "../types";
 import {
   getTokenById,
   findCompoundToken,
-  findTokenByAddress,
-  findTokenById,
   formatCurrencyUnit,
 } from "../../../currencies";
 import network from "../../../network";
 import { promiseAllBatched } from "../../../promise";
 import { mergeOps } from "../../../bridge/jsHelpers";
 import { apiForCurrency } from "../../../api/Ethereum";
-import type { Transaction } from "../types";
 import { inferTokenAccount } from "../transaction";
 
 const compoundWhitelist = [
@@ -49,15 +46,11 @@ export function isCompoundTokenSupported(token: TokenCurrency): boolean {
 
 export type Modes = "compound.supply" | "compound.withdraw";
 
-export function contractField(transaction: Transaction) {
-  const recipientToken = findTokenByAddress(transaction.recipient);
-  const maybeCompoundToken = findTokenById(recipientToken?.compoundFor || "");
+function contractField(ctoken) {
   return {
     type: "text",
-    label: maybeCompoundToken ? "Contract" : "Address",
-    value: maybeCompoundToken
-      ? "Compound " + maybeCompoundToken.ticker
-      : transaction.recipient,
+    label: "Contract",
+    value: "Compound " + ctoken.ticker,
   };
 }
 
@@ -110,7 +103,7 @@ const compoundSupply: ModeModule = {
       }),
     });
 
-    fields.push(contractField(transaction));
+    fields.push(contractField(ctoken));
   },
   fillOptimisticOperation(a, t, op) {
     const subAccount = inferTokenAccount(a, t);
@@ -131,7 +124,7 @@ const compoundSupply: ModeModule = {
           blockHash: null,
           blockHeight: null,
           senders: op.senders,
-          recipients: [t.recipient],
+          recipients: op.recipients,
           accountId: subAccount.id,
           date: new Date(),
           extra: {
@@ -218,7 +211,7 @@ const compoundWithdraw: ModeModule = {
       value,
     });
 
-    fields.push(contractField(transaction));
+    fields.push(contractField(ctoken));
   },
   fillOptimisticOperation(a, t, op) {
     const subAccount = inferTokenAccount(a, t);
@@ -244,7 +237,7 @@ const compoundWithdraw: ModeModule = {
           blockHash: null,
           blockHeight: null,
           senders: op.senders,
-          recipients: [t.recipient],
+          recipients: op.recipients,
           accountId: subAccount.id,
           date: new Date(),
           extra: {
