+++ src/families/ethereum/modules/erc20.js
@@ -9,7 +9,19 @@ import type { ModeModule } from "../types";
 import { AmountRequired } from "@ledgerhq/errors";
 import { inferTokenAccount, validateRecipient } from "../transaction";
 import { getAccountCurrency } from "../../../account";
-import { contractField } from "./compound";
+import { findTokenByAddress, findTokenById } from "../../../currencies";
+
+function contractField(transaction) {
+  const recipientToken = findTokenByAddress(transaction.recipient);
+  const maybeCompoundToken = findTokenById(recipientToken?.compoundFor || "");
+  return {
+    type: "text",
+    label: maybeCompoundToken ? "Contract" : "Address",
+    value: maybeCompoundToken
+      ? "Compound " + maybeCompoundToken.ticker
+      : transaction.recipient,
+  };
+}
 
 export type Modes = "erc20.approve";
